requirements.txt
Django>=5.1
djangorestframework
channels
channels-redis
uvicorn
django-cors-headers
psycopg2-binary==2.9.8